♻️ Refactor file upload
- moved upload function to utils
- fetch files after uploading
- handle upload cleanly with Promise.all()
