+++ client/src/utils/index.ts
@@ -32,7 +32,7 @@ export async function getFiles(): Promise<DisplayFile[]> {
   return res.data.data;
 }
 
-export async function upload(
+export async function uploadService(
   data: any,
   options: {
     onUploadProgress?: (progressEvent: any) => void;
@@ -44,4 +44,49 @@ export async function upload(
   };
   const res = await axios.post(`${API_ORIGIN}/api/files`, data, options);
   return res.data;
+}
+
+export async function uploadFiles(
+  evt: Event,
+  cb: (num: number | null) => void
+) {
+  const evtTarget = evt.target as HTMLInputElement;
+
+  if (!evtTarget.files) return;
+  if (evtTarget.files.length < 0) return;
+
+  const uploadTracker: Record<string, number> = {};
+
+  const files = Array.from(evtTarget.files);
+  console.log("upload files", files);
+
+  const uploadFilesPromises = files.map(async (file: File) => {
+    let formDataFile = new FormData();
+    formDataFile.append("upload", file);
+    uploadTracker[file.name] = 0;
+
+    return uploadService(formDataFile, {
+      onUploadProgress: (event: ProgressEvent) => {
+        if (!event.lengthComputable) return;
+
+        const uploadProgress: number = (event.loaded * 100) / event.total;
+        uploadTracker[file.name] = Math.round(uploadProgress);
+
+        const scores: number[] = Object.values(uploadTracker);
+        const percentCompleted = Math.floor(
+          scores.reduce((prevScore: number, currentScore: number) => {
+            return prevScore + currentScore;
+          }, 0) / scores.length
+        );
+
+        cb(percentCompleted);
+        // console.log(percentCompleted, scores, scores.length);
+      },
+    });
+  });
+
+  await Promise.all(uploadFilesPromises);
+
+  alert("Upload complete");
+  cb(null);
 }
